+++ ios/Classes/FLTInstanceManager_Test.h
@@ -2,11 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import "InstanceManager.h"
+#import "FLTInstanceManager.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface InstanceManager ()
+@interface FLTInstanceManager ()
 /// The next identifier that will be used for a host-created instance.
 @property long nextIdentifier;