+++ lib/flutter_instance_manager.dart
@@ -1,3 +1,3 @@
 library flutter_instance_manager;
 
-export 'src/shared_instance_manager.dart';
+export 'src/instance_manager.dart';