+++ test/instance_manager_test.dart
@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license.
 
-import 'package:flutter_instance_manager/src/shared_instance_manager.dart';
+import 'package:flutter_instance_manager/flutter_instance_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:uuid/uuid.dart';