+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.2.0
 
-* Adds iOS support
+* Adds iOS support for the instance manager.
 
 # 0.1.0
 
-* Adds Android support
+* Adds Android support for the instance manager.